store patch radius in pixels in SkyPatch in initialization
SkyPatch is mainly (other than in tests) initialized from a
CatalogEntry, where a radius in pixels is calculated.  Formerly, that
radius was immediately converted to world coordinates and stored,
discarding the radius in pixels. However, we use the radius in pixels
all over the place, so it makes sense to store it instead of converting
back to pixels all the time.


Former-commit-id: e97fc2856628436b906dd38d5ffe6e5a3862fcd7