generate_test_image.py: simplify psf; generate images from truth csv
First, I've removed all the fancy PSF and other nonsense that isn't really
important at this stage.

Second, I've changed this script around. Instead of writing a grid of randomized
images along with a truth CSV, it now *reads* a list of test cases (in the form
of "true" galaxy and noise parameters) from a truth CSV, and writes a series of
images, one for each row. The julia script can then read in this CSV, read the
corresponding image for each row, run Celeste, and compare its output to the CSV
values.