update GalsimBenchmark to read truth CSV and run on corresponding series of images
In theory this can run on each image and report the inferred vs. "true" values, as well as comparing
them and reporting some measure of accuracy (or flag serious errors). In practice I haven't
gotten there yet because the optimization routine is crashing and I need to learn how to debug and
fix that before I can move ahead with the benchmark scaffolding.