Throw exception if any value of epsilon_mat == 0 (and add Exceptions module)
A zero value of epsilon_mat can lead to expected pixel brightness (elbo_vars.E_G) going to zero,
which causes the ELBO evaluation to fail.

We could, say, reuse the built-in DomainError, but then we have no opportunity to communicate to the
user what exactly was wrong. Even in the simple case of sqrt(-5) this was a serious deficiency for
which the julia developers put in a [ridiculous
hack](https://github.com/JuliaLang/julia/blob/0baa867307e0c59e154d5b638b44da600dd2d636/base/replutil.jl#L225).

I don't know if a dedicated module like this is a common pattern in Julia, but I've done it a lot in
Python, so let me know if it seems bad.