Refactor to define GalSim cases in Python instead of CSV
The previous design defined each test case with a row of a CSV file. This was a simple solution to
the problem of sharing "ground truth" information between the Python script that generates images
and the Julia code that runs Celeste.

Turns out it's more elegant to put the ground truth information in the FITS header: this keeps all
the relevant test case information in a single file and makes it possible for other people to grab
the FITS file for their own accuracy benchmarks.

With that in place there's no need for the CSV, and describing the test cases in Python is nicer. It
lets us share defaults and reuse one test case in another, reducing duplication. And it makes it
possible to describe more complex test cases such as multiple overlapping light sources.

The code here may feel overdesigned and let me know if you think so. My goal was to avoid
duplication and make the test case definitions as easy and readable as possible (without resorting
to crazy Python magic).