expand image + shrink galaxies to avoid clipping
Turns out this was broken since I expanded the PSF sigma, but the
`assert_counts_match_expected_flux()` wasn't catching it because `CL_NOISE` was erroneously set to
true in all images' FITS headers. The refactoring to describe test cases in Python fixed that,
exposing this problem via the assertion.