galsim test cases w/ different PSFs
This changes the PSF sigma from a global constant in test_case_definitions.py to a per-test-case
variable. Correspondingly, we write the PSF sigma into the FITS header. I added two test cases with
a PSF smaller and larger than the usual one, so we can test what happens when Celeste runs with a
mis-specified PSF.

Fixes #388