switch to photon shooting for galsim generation
this is much, much faster for the 500-source field. (the default FFT method didn't even work when I
added de Vaucouleurs profiles, I think because they decay more slowly, so required larger FFTs.) i
think the tradeoff is that it's inexact, but i've set it to compute within 0.1% of the sky noise
level and it's plenty fast, so seems ok to me, but we'll have to monitor for weirdness.