Fully support `e_dev` in GalSim benchmarks (#539)
* support e_dev in galsim generation lib; add benchmark cases

* report e_dev expected/actual in GalsimBenchmark; udpate unit test error threshold

* generate and use e_dev in galsim 500-source field

* set seed in 500-source field catalog generation

* show per-source progress in galsim generator

the 500-source field can be slow and this progress display is useful

* switch to photon shooting for galsim generation

this is much, much faster for the 500-source field. (the default FFT method didn't even work when I
added de Vaucouleurs profiles, I think because they decay more slowly, so required larger FFTs.) i
think the tradeoff is that it's inexact, but i've set it to compute within 0.1% of the sky noise
level and it's plenty fast, so seems ok to me, but we'll have to monitor for weirdness.