move load_primary into AccuracyBenchmark + refactor
the refactoring makes it resemble `load_coadd_catalog`, but it would be annoying to make them
actually share code, since a bunch of field names differ and some of the formats differ slightly. i
think the thing to do might be to have a script that pulls primary catalog and imagery through
SDSSIO into a standalone catalog and image file which matches the format of the synthetic data. in
which case this problem will go away. but then we're juggling the ground truth catalog, the
"initialization" catalog and the imagery for all these cases, and we probably ought to just embed
everything into a single FITS file with mixed table and image HDUs. and i'm putting that off for now
to avoid feature creep.