clean up some imports in tests
These test files aren't under the Celeste module so they shouldn't really be doing .. imports, I
think. It works because they're included into the runtests.jl script which imports a bunch of stuff,
but it seems confusing, prone to breakage, and it doesn't work when you pull code out into sandbox
scripts (for use with `reload()`).