Fixed position offset in accuracy benchmarks (#583)
* fix bug where i forgot to add the initial position offset

* initialize with a fixed, not random, position offset

random seems preferable for broader test coverage, but inference can be very sensitive to this, as
on the `simple_star` test case: we infer probability of star = .005 when initializing with exactly
the true position, and 0.995 when initialized slightly offset from the true position. (that's
problematic by itself, but it's a different discussion.) since inference is sensitive, let's not
randomize this offset and risk having extremely brittle tests (i.e., even if we srand, if some
computations are reordered the random offset would change and the test could mysteriously break).