support generating imagery from catalog CSV using Synthetic.jl
The is the analog of benchmark/galsim/galsim_field.py, using Synthetic.jl for drawing pixels from
the Celeste likelihood model. Unfortunately a small amount of logic needs to be duplicated between
Python and Julia.

I think there's an issue with e_angle still, but i'll work on it next.