Refactor IO strategies (#670)
This adds some more flexibity to I/O, for non-NERSC environemnts.
In particular it allows:
- Data in the original SDSS directory layout, rather than the one that Celeste puts them in
- Decompresing data on the fly rather than requiring it to be decompressed on disk
- Clusters without shared filesystems by doing RPC to a master process

Since the options were getting a bit out of hand with these additions,
I also added a YAML based configuration format to specify these options.
I plan to extend that settings file to cover other aspects of the system
as well (e.g. the distributed options).

Finally, I removed the big file option. That was always a bit of a stopgap.
We didn't document the format. The scripts to create them are not in tree.
With more time, we can probably come up with a better way to do it. Also,
it's untested. It's still in the history of course, so we can always bring
it back.

Lastly, I added tests for all the new I/O modes.