Isolate Gasp-dependent code
There's been a long standing usability problem with Celeste, namely that the value of
the USE_DTREE environment variable at precompile-time would affect whether or not
distributed processing was available. This patch aims to address that by factoring
multinode_run into its own module that can be independently loaded over a base copy
of Celeste, making it possible to load Gasp-dependent functionality on demand.

To maintain the user interface the bin/ script still respects the environment variable
for now, but in the future this could be handled via settings file or command line
arguments.