Move NERSC scripts to new contrib/ folder
`contrib` is a common name for utilities in using the software that are not part of
the main software, but useful for using it in certain environments. The nersc/
folder is essentially that for the NERSC environment, but I'm planning to add a couple
of extra utilities, so it would be good to have a proper place for these.