Grabbag of local changes (#673)
* Switch puts for an explicit write

When julia redirects output, it redirects the FD, but not the C standard streams.
This makes sure that the output still shows up in this case.

* Integrate FieldExtents into the new IO setup

* Isolate Gasp-dependent code

There's been a long standing usability problem with Celeste, namely that the value of
the USE_DTREE environment variable at precompile-time would affect whether or not
distributed processing was available. This patch aims to address that by factoring
multinode_run into its own module that can be independently loaded over a base copy
of Celeste, making it possible to load Gasp-dependent functionality on demand.

To maintain the user interface the bin/ script still respects the environment variable
for now, but in the future this could be handled via settings file or command line
arguments.

* Move NERSC scripts to new contrib/ folder

`contrib` is a common name for utilities in using the software that are not part of
the main software, but useful for using it in certain environments. The nersc/
folder is essentially that for the NERSC environment, but I'm planning to add a couple
of extra utilities, so it would be good to have a proper place for these.

* Add contrib code for running Celeste over HTTP

This is useful for running Celeste on systems without a shared filesystem,
and more generally shows how to extent Celeste to a new IO strategy.